# openapi.yml
openapi: 3.0.0
info:
  title: Get Spotify Data
  version: 1.0.0

paths:
  /api/playlist/{playlist_id}:
    get:
      summary: Get Playlist Metadata + Tracks
      operationId: getPlaylist
      parameters:
        - $ref: "#/components/parameters/AuthSecret"
        - name: playlist_id
          in: path
          required: true
          schema:
            type: string

      responses:
        "200":
          description: Playlist data with tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/artist/{artist_id}:
    get:
      summary: Get Arist Metadata
      operationId: getArtist
      parameters:
        - $ref: "#/components/parameters/AuthSecret"
        - name: artist_id
          in: path
          required: true
          schema:
            type: string

      responses:
        "200":
          description: Artist data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  parameters:
    AuthSecret:
      name: secret
      in: query
      required: true
      schema:
        type: string

  responses:
    Unauthorized:
      description: Unauthorized — secret is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Token:
      type: object
      properties:
        access_token:
          type: string
          description: The access token to use for API calls.
        expiry:
          type: integer
          description: The UNIX timestamp when the token expires.
      required:
        - access_token
        - expiry

    Error:
      type: object
      properties:
        error:
          type: string
        display_message:
          type: string
      required:
        - error

    Image:
      type: object
      properties:
        url:
          type: string
        height:
          type: integer
        width:
          type: integer
      required:
        - url
        - height
        - width

    SimplifiedArtist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name

    User:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
      required:
        - id
        - display_name

    PlaylistTrack:
      type: object
      properties:
        added_at:
          type: string
        is_local:
          type: boolean
        track:
          $ref: "#/components/schemas/Track"
      required:
        - added_at
        - is_local
        - track

    Track:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        artists:
          type: array
          items:
            $ref: "#/components/schemas/SimplifiedArtist"
        duration_ms:
          type: integer
      required:
        - id
        - name
        - artists
        - duration_ms

    TrackPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlaylistTrack"
      required:
        - items

    Playlist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        owner:
          $ref: "#/components/schemas/User"
        tracks:
          $ref: "#/components/schemas/TrackPage"
      required:
        - id
        - name
        - description
        - images
        - owner
        - tracks

    Artist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
      required:
        - id
        - name
        - description
        - images
        - owner
        - tracks